Test arithmetic functions at the DSL level
